Update event handling code for compatibility with jQuery 3.0 and above
This commit updates the event handling code to ensure compatibility
with jQuery 3.0 and higher versions. The previous code relied on the
window.jQuery.event.props property, which has been removed in jQuery 3.0.

The updated code checks if the dataTransfer property is defined in the
jQuery.event.special object. If not, it adds the property to ensure
proper functionality. This modification allows the code to work seamlessly
with jQuery 3.0 and above. For more info. please refer [1] and [2].

[1] https://jquery.com/upgrade-guide/3.0/#breaking-change-jquery-event-props-and-jquery-event-fixhooks-removed
[2] https://learn.jquery.com/events/event-extensions/

Closes-Bug: 2027728

Change-Id: Ibb6db39b3e78eb6634f06a10417a76d529703de8